+++ src/modules/financials/entities/bi-notul.ts
@@ -35,4 +35,8 @@ export class Notul extends BaseEntity {
   @Field()
   @Column({ name: 'IMPT_ID' })
   imptId!: number;
+
+  @Field({ nullable: true })
+  @Column({ name: 'CATEGORY' })
+  category?: string;
 }